Add refatored trips and expenses class, ready for using in calc functions

+++ src/SerializedDataYears.java
@@ -1,11 +1,13 @@
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Agnieszka on 29.12.2016.
  */
 public class SerializedDataYears {
     @SerializedName("pola")
-    Double[] pola;
+    BigDecimal[] pola;
 
     @SerializedName("dokument_id")
     int dokument_id;
